+++ src/Api/Billing/Controllers/OrganizationBillingController.cs
@@ -1,5 +1,11 @@
 ﻿using Bit.Api.Billing.Models.Responses;
+using Bit.Api.Models.Response;
 using Bit.Core.Billing.Queries;
+using Bit.Core.Context;
+using Bit.Core.Exceptions;
+using Bit.Core.Repositories;
+using Bit.Core.Services;
+using Bit.Core.Utilities;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -8,7 +14,10 @@ namespace Bit.Api.Billing.Controllers;
 [Route("organizations/{organizationId:guid}/billing")]
 [Authorize("Application")]
 public class OrganizationBillingController(
-    IOrganizationBillingQueries organizationBillingQueries) : Controller
+    IOrganizationBillingQueries organizationBillingQueries,
+    ICurrentContext currentContext,
+    IOrganizationRepository organizationRepository,
+    IPaymentService paymentService) : Controller
 {
     [HttpGet("metadata")]
     public async Task<IResult> GetMetadataAsync([FromRoute] Guid organizationId)
@@ -23,5 +32,25 @@ public class OrganizationBillingController(
         var response = OrganizationMetadataResponse.From(metadata);
 
         return TypedResults.Ok(response);
+    }
+
+    [HttpGet]
+    [SelfHosted(NotSelfHostedOnly = true)]
+    public async Task<BillingResponseModel> GetBilling(string id)
+    {
+        var orgIdGuid = new Guid(id);
+        if (!await currentContext.ViewBillingHistory(orgIdGuid))
+        {
+            throw new NotFoundException();
+        }
+
+        var organization = await organizationRepository.GetByIdAsync(orgIdGuid);
+        if (organization == null)
+        {
+            throw new NotFoundException();
+        }
+
+        var billingInfo = await paymentService.GetBillingAsync(organization);
+        return new BillingResponseModel(billingInfo);
     }
 }
