+++ test/Api.Test/AdminConsole/Controllers/OrganizationsControllerTests.cs
@@ -1,9 +1,7 @@
 ﻿using System.Security.Claims;
 using AutoFixture.Xunit2;
 using Bit.Api.AdminConsole.Controllers;
-using Bit.Api.AdminConsole.Models.Request.Organizations;
 using Bit.Api.Auth.Models.Request.Accounts;
-using Bit.Api.Models.Request.Organizations;
 using Bit.Core;
 using Bit.Core.AdminConsole.Entities;
 using Bit.Core.AdminConsole.Enums.Provider;
@@ -16,21 +14,14 @@ using Bit.Core.Auth.Models.Data;
 using Bit.Core.Auth.Repositories;
 using Bit.Core.Auth.Services;
 using Bit.Core.Billing.Commands;
-using Bit.Core.Billing.Queries;
 using Bit.Core.Context;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
 using Bit.Core.Exceptions;
-using Bit.Core.Models.Business;
-using Bit.Core.Models.Data.Organizations.OrganizationUsers;
-using Bit.Core.OrganizationFeatures.OrganizationLicenses.Interfaces;
-using Bit.Core.OrganizationFeatures.OrganizationSubscriptions.Interface;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
-using Bit.Core.Tools.Services;
 using Bit.Infrastructure.EntityFramework.AdminConsole.Models.Provider;
 using NSubstitute;
-using NSubstitute.ReturnsExtensions;
 using Xunit;
 using GlobalSettings = Bit.Core.Settings.GlobalSettings;
 
@@ -43,7 +34,6 @@ public class OrganizationsControllerTests : IDisposable
     private readonly IOrganizationRepository _organizationRepository;
     private readonly IOrganizationService _organizationService;
     private readonly IOrganizationUserRepository _organizationUserRepository;
-    private readonly IPaymentService _paymentService;
     private readonly IPolicyRepository _policyRepository;
     private readonly ISsoConfigRepository _ssoConfigRepository;
     private readonly ISsoConfigService _ssoConfigService;
@@ -51,17 +41,9 @@ public class OrganizationsControllerTests : IDisposable
     private readonly IGetOrganizationApiKeyQuery _getOrganizationApiKeyQuery;
     private readonly IRotateOrganizationApiKeyCommand _rotateOrganizationApiKeyCommand;
     private readonly IOrganizationApiKeyRepository _organizationApiKeyRepository;
-    private readonly ICloudGetOrganizationLicenseQuery _cloudGetOrganizationLicenseQuery;
     private readonly ICreateOrganizationApiKeyCommand _createOrganizationApiKeyCommand;
     private readonly IFeatureService _featureService;
-    private readonly ILicensingService _licensingService;
-    private readonly IUpdateSecretsManagerSubscriptionCommand _updateSecretsManagerSubscriptionCommand;
-    private readonly IUpgradeOrganizationPlanCommand _upgradeOrganizationPlanCommand;
-    private readonly IAddSecretsManagerSubscriptionCommand _addSecretsManagerSubscriptionCommand;
     private readonly IPushNotificationService _pushNotificationService;
-    private readonly ICancelSubscriptionCommand _cancelSubscriptionCommand;
-    private readonly ISubscriberQueries _subscriberQueries;
-    private readonly IReferenceEventService _referenceEventService;
     private readonly IOrganizationEnableCollectionEnhancementsCommand _organizationEnableCollectionEnhancementsCommand;
     private readonly IProviderRepository _providerRepository;
     private readonly IScaleSeatsCommand _scaleSeatsCommand;
@@ -75,7 +57,6 @@ public class OrganizationsControllerTests : IDisposable
         _organizationRepository = Substitute.For<IOrganizationRepository>();
         _organizationService = Substitute.For<IOrganizationService>();
         _organizationUserRepository = Substitute.For<IOrganizationUserRepository>();
-        _paymentService = Substitute.For<IPaymentService>();
         _policyRepository = Substitute.For<IPolicyRepository>();
         _ssoConfigRepository = Substitute.For<ISsoConfigRepository>();
         _ssoConfigService = Substitute.For<ISsoConfigService>();
@@ -83,17 +64,9 @@ public class OrganizationsControllerTests : IDisposable
         _rotateOrganizationApiKeyCommand = Substitute.For<IRotateOrganizationApiKeyCommand>();
         _organizationApiKeyRepository = Substitute.For<IOrganizationApiKeyRepository>();
         _userService = Substitute.For<IUserService>();
-        _cloudGetOrganizationLicenseQuery = Substitute.For<ICloudGetOrganizationLicenseQuery>();
         _createOrganizationApiKeyCommand = Substitute.For<ICreateOrganizationApiKeyCommand>();
         _featureService = Substitute.For<IFeatureService>();
-        _licensingService = Substitute.For<ILicensingService>();
-        _updateSecretsManagerSubscriptionCommand = Substitute.For<IUpdateSecretsManagerSubscriptionCommand>();
-        _upgradeOrganizationPlanCommand = Substitute.For<IUpgradeOrganizationPlanCommand>();
-        _addSecretsManagerSubscriptionCommand = Substitute.For<IAddSecretsManagerSubscriptionCommand>();
         _pushNotificationService = Substitute.For<IPushNotificationService>();
-        _cancelSubscriptionCommand = Substitute.For<ICancelSubscriptionCommand>();
-        _subscriberQueries = Substitute.For<ISubscriberQueries>();
-        _referenceEventService = Substitute.For<IReferenceEventService>();
         _organizationEnableCollectionEnhancementsCommand = Substitute.For<IOrganizationEnableCollectionEnhancementsCommand>();
         _providerRepository = Substitute.For<IProviderRepository>();
         _scaleSeatsCommand = Substitute.For<IScaleSeatsCommand>();
@@ -104,7 +77,6 @@ public class OrganizationsControllerTests : IDisposable
             _policyRepository,
             _organizationService,
             _userService,
-            _paymentService,
             _currentContext,
             _ssoConfigRepository,
             _ssoConfigService,
@@ -112,17 +84,9 @@ public class OrganizationsControllerTests : IDisposable
             _rotateOrganizationApiKeyCommand,
             _createOrganizationApiKeyCommand,
             _organizationApiKeyRepository,
-            _cloudGetOrganizationLicenseQuery,
             _featureService,
             _globalSettings,
-            _licensingService,
-            _updateSecretsManagerSubscriptionCommand,
-            _upgradeOrganizationPlanCommand,
-            _addSecretsManagerSubscriptionCommand,
             _pushNotificationService,
-            _cancelSubscriptionCommand,
-            _subscriberQueries,
-            _referenceEventService,
             _organizationEnableCollectionEnhancementsCommand,
             _providerRepository,
             _scaleSeatsCommand);
@@ -191,196 +155,6 @@ public class OrganizationsControllerTests : IDisposable
 
         await _organizationService.DeleteUserAsync(orgId, user.Id);
         await _organizationService.Received(1).DeleteUserAsync(orgId, user.Id);
-    }
-
-    [Theory, AutoData]
-    public async Task OrganizationsController_PostUpgrade_UserCannotEditSubscription_ThrowsNotFoundException(
-        Guid organizationId,
-        OrganizationUpgradeRequestModel model)
-    {
-        _currentContext.EditSubscription(organizationId).Returns(false);
-
-        await Assert.ThrowsAsync<NotFoundException>(() => _sut.PostUpgrade(organizationId.ToString(), model));
-    }
-
-    [Theory, AutoData]
-    public async Task OrganizationsController_PostUpgrade_NonSMUpgrade_ReturnsCorrectResponse(
-        Guid organizationId,
-        OrganizationUpgradeRequestModel model,
-        bool success,
-        string paymentIntentClientSecret)
-    {
-        model.UseSecretsManager = false;
-
-        _currentContext.EditSubscription(organizationId).Returns(true);
-
-        _upgradeOrganizationPlanCommand.UpgradePlanAsync(organizationId, Arg.Any<OrganizationUpgrade>())
-            .Returns(new Tuple<bool, string>(success, paymentIntentClientSecret));
-
-        var response = await _sut.PostUpgrade(organizationId.ToString(), model);
-
-        Assert.Equal(success, response.Success);
-        Assert.Equal(paymentIntentClientSecret, response.PaymentIntentClientSecret);
-    }
-
-    [Theory, AutoData]
-    public async Task OrganizationsController_PostUpgrade_SMUpgrade_ProvidesAccess_ReturnsCorrectResponse(
-        Guid organizationId,
-        Guid userId,
-        OrganizationUpgradeRequestModel model,
-        bool success,
-        string paymentIntentClientSecret,
-        OrganizationUser organizationUser)
-    {
-        model.UseSecretsManager = true;
-        organizationUser.AccessSecretsManager = false;
-
-        _currentContext.EditSubscription(organizationId).Returns(true);
-
-        _upgradeOrganizationPlanCommand.UpgradePlanAsync(organizationId, Arg.Any<OrganizationUpgrade>())
-            .Returns(new Tuple<bool, string>(success, paymentIntentClientSecret));
-
-        _userService.GetProperUserId(Arg.Any<ClaimsPrincipal>()).Returns(userId);
-
-        _organizationUserRepository.GetByOrganizationAsync(organizationId, userId).Returns(organizationUser);
-
-        var response = await _sut.PostUpgrade(organizationId.ToString(), model);
-
-        Assert.Equal(success, response.Success);
-        Assert.Equal(paymentIntentClientSecret, response.PaymentIntentClientSecret);
-
-        await _organizationUserRepository.Received(1).ReplaceAsync(Arg.Is<OrganizationUser>(orgUser =>
-            orgUser.Id == organizationUser.Id && orgUser.AccessSecretsManager == true));
-    }
-
-    [Theory, AutoData]
-    public async Task OrganizationsController_PostUpgrade_SMUpgrade_NullOrgUser_ReturnsCorrectResponse(
-        Guid organizationId,
-        Guid userId,
-        OrganizationUpgradeRequestModel model,
-        bool success,
-        string paymentIntentClientSecret)
-    {
-        model.UseSecretsManager = true;
-
-        _currentContext.EditSubscription(organizationId).Returns(true);
-
-        _upgradeOrganizationPlanCommand.UpgradePlanAsync(organizationId, Arg.Any<OrganizationUpgrade>())
-            .Returns(new Tuple<bool, string>(success, paymentIntentClientSecret));
-
-        _userService.GetProperUserId(Arg.Any<ClaimsPrincipal>()).Returns(userId);
-
-        _organizationUserRepository.GetByOrganizationAsync(organizationId, userId).ReturnsNull();
-
-        var response = await _sut.PostUpgrade(organizationId.ToString(), model);
-
-        Assert.Equal(success, response.Success);
-        Assert.Equal(paymentIntentClientSecret, response.PaymentIntentClientSecret);
-
-        await _organizationUserRepository.DidNotReceiveWithAnyArgs().ReplaceAsync(Arg.Any<OrganizationUser>());
-    }
-
-    [Theory, AutoData]
-    public async Task OrganizationsController_PostSubscribeSecretsManagerAsync_NullOrg_ThrowsNotFoundException(
-        Guid organizationId,
-        SecretsManagerSubscribeRequestModel model)
-    {
-        _organizationRepository.GetByIdAsync(organizationId).ReturnsNull();
-
-        await Assert.ThrowsAsync<NotFoundException>(() => _sut.PostSubscribeSecretsManagerAsync(organizationId, model));
-    }
-
-    [Theory, AutoData]
-    public async Task OrganizationsController_PostSubscribeSecretsManagerAsync_UserCannotEditSubscription_ThrowsNotFoundException(
-        Guid organizationId,
-        SecretsManagerSubscribeRequestModel model,
-        Organization organization)
-    {
-        _organizationRepository.GetByIdAsync(organizationId).Returns(organization);
-
-        _currentContext.EditSubscription(organizationId).Returns(false);
-
-        await Assert.ThrowsAsync<NotFoundException>(() => _sut.PostSubscribeSecretsManagerAsync(organizationId, model));
-    }
-
-    [Theory, AutoData]
-    public async Task OrganizationsController_PostSubscribeSecretsManagerAsync_ProvidesAccess_ReturnsCorrectResponse(
-        Guid organizationId,
-        SecretsManagerSubscribeRequestModel model,
-        Organization organization,
-        Guid userId,
-        OrganizationUser organizationUser,
-        OrganizationUserOrganizationDetails organizationUserOrganizationDetails)
-    {
-        organizationUser.AccessSecretsManager = false;
-
-        var ssoConfigurationData = new SsoConfigurationData
-        {
-            MemberDecryptionType = MemberDecryptionType.KeyConnector,
-            KeyConnectorUrl = "https://example.com"
-        };
-
-        organizationUserOrganizationDetails.Permissions = string.Empty;
-        organizationUserOrganizationDetails.SsoConfig = ssoConfigurationData.Serialize();
-
-        _organizationRepository.GetByIdAsync(organizationId).Returns(organization);
-
-        _currentContext.EditSubscription(organizationId).Returns(true);
-
-        _userService.GetProperUserId(Arg.Any<ClaimsPrincipal>()).Returns(userId);
-
-        _organizationUserRepository.GetByOrganizationAsync(organization.Id, userId).Returns(organizationUser);
-
-        _organizationUserRepository.GetDetailsByUserAsync(userId, organization.Id, OrganizationUserStatusType.Confirmed)
-            .Returns(organizationUserOrganizationDetails);
-
-        var response = await _sut.PostSubscribeSecretsManagerAsync(organizationId, model);
-
-        Assert.Equal(response.Id, organizationUserOrganizationDetails.OrganizationId);
-        Assert.Equal(response.Name, organizationUserOrganizationDetails.Name);
-
-        await _addSecretsManagerSubscriptionCommand.Received(1)
-            .SignUpAsync(organization, model.AdditionalSmSeats, model.AdditionalServiceAccounts);
-        await _organizationUserRepository.Received(1).ReplaceAsync(Arg.Is<OrganizationUser>(orgUser =>
-            orgUser.Id == organizationUser.Id && orgUser.AccessSecretsManager == true));
-    }
-
-    [Theory, AutoData]
-    public async Task OrganizationsController_PostSubscribeSecretsManagerAsync_NullOrgUser_ReturnsCorrectResponse(
-        Guid organizationId,
-        SecretsManagerSubscribeRequestModel model,
-        Organization organization,
-        Guid userId,
-        OrganizationUserOrganizationDetails organizationUserOrganizationDetails)
-    {
-        var ssoConfigurationData = new SsoConfigurationData
-        {
-            MemberDecryptionType = MemberDecryptionType.KeyConnector,
-            KeyConnectorUrl = "https://example.com"
-        };
-
-        organizationUserOrganizationDetails.Permissions = string.Empty;
-        organizationUserOrganizationDetails.SsoConfig = ssoConfigurationData.Serialize();
-
-        _organizationRepository.GetByIdAsync(organizationId).Returns(organization);
-
-        _currentContext.EditSubscription(organizationId).Returns(true);
-
-        _userService.GetProperUserId(Arg.Any<ClaimsPrincipal>()).Returns(userId);
-
-        _organizationUserRepository.GetByOrganizationAsync(organization.Id, userId).ReturnsNull();
-
-        _organizationUserRepository.GetDetailsByUserAsync(userId, organization.Id, OrganizationUserStatusType.Confirmed)
-            .Returns(organizationUserOrganizationDetails);
-
-        var response = await _sut.PostSubscribeSecretsManagerAsync(organizationId, model);
-
-        Assert.Equal(response.Id, organizationUserOrganizationDetails.OrganizationId);
-        Assert.Equal(response.Name, organizationUserOrganizationDetails.Name);
-
-        await _addSecretsManagerSubscriptionCommand.Received(1)
-            .SignUpAsync(organization, model.AdditionalSmSeats, model.AdditionalServiceAccounts);
-        await _organizationUserRepository.DidNotReceiveWithAnyArgs().ReplaceAsync(Arg.Any<OrganizationUser>());
     }
 
     [Theory, AutoData]
